backend/requirements_prod.txt
httplib2==0.22.0
python-dotenv==1.0.1
flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-bcrypt==1.0.1
flask-restx==1.3.0
flask-jwt-extended==4.6.0
Flask-Cors==5.0.0
psycopg2-binary==2.9.3
boto3==1.35.82
flask-socketio==5.4.1
stripe==11.4.1
pandas==2.2.3
nltk==3.9.1
gensim==4.3.3
scikit-learn==1.6.1
Flask-Mail==0.10.0
pillow==11.1.0
qrcode==8.0
implicit==0.7.2